Yet another micro optimisation: we avoid string copies this way - all keys in our JSON objects are static strings, so we can afford using simply std::string_view.



<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 10515.6<br/>plain u32: 10444.7<br/>aliased double: 14342.1<br/>plain double: 14304.7 | aliased u32: 10988.2<br/>plain u32: 10917.5<br/>aliased double: 15052.9<br/>plain double: 15049.8 |
| e2e_match_ch | Ops: 25.85 ± 0.03 ops/s. Best: 25.88 ops/s<br/>Total: 5068.11ms ± 5.67ms. Best: 5061.04ms<br/>Min time: 2.96ms ± 0.07ms<br/>Mean time: 38.69ms ± 0.04ms<br/>Median time: 30.41ms ± 0.22ms<br/>95th percentile: 120.62ms ± 0.43ms<br/>99th percentile: 150.63ms ± 0.67ms<br/>Max time: 174.53ms ± 0.75ms | Ops: 25.47 ± 0.01 ops/s. Best: 25.48 ops/s<br/>Total: 5143.03ms ± 2.67ms. Best: 5140.79ms<br/>Min time: 2.93ms ± 0.05ms<br/>Mean time: 39.26ms ± 0.02ms<br/>Median time: 30.74ms ± 0.13ms<br/>95th percentile: 121.65ms ± 0.44ms<br/>99th percentile: 154.04ms ± 0.23ms<br/>Max time: 177.92ms ± 0.57ms |
| e2e_match_mld | Ops: 43.27 ± 0.04 ops/s. Best: 43.32 ops/s<br/>Total: 3027.27ms ± 2.42ms. Best: 3023.94ms<br/>Min time: 2.41ms ± 0.03ms<br/>Mean time: 23.11ms ± 0.02ms<br/>Median time: 12.54ms ± 0.05ms<br/>95th percentile: 75.41ms ± 0.22ms<br/>99th percentile: 87.46ms ± 0.33ms<br/>Max time: 100.47ms ± 0.37ms | Ops: 43.09 ± 0.01 ops/s. Best: 43.12 ops/s<br/>Total: 3040.10ms ± 0.84ms. Best: 3038.14ms<br/>Min time: 2.35ms ± 0.02ms<br/>Mean time: 23.21ms ± 0.01ms<br/>Median time: 12.57ms ± 0.04ms<br/>95th percentile: 75.73ms ± 0.13ms<br/>99th percentile: 87.80ms ± 0.24ms<br/>Max time: 101.27ms ± 0.37ms |
| e2e_nearest_ch | Ops: 672.02 ± 3.71 ops/s. Best: 678.52 ops/s<br/>Total: 1488.05ms ± 8.20ms. Best: 1473.81ms<br/>Min time: 1.22ms ± 0.01ms<br/>Mean time: 1.49ms ± 0.01ms<br/>Median time: 1.45ms ± 0.01ms<br/>95th percentile: 1.85ms ± 0.01ms<br/>99th percentile: 1.93ms ± 0.02ms<br/>Max time: 8.91ms ± 7.09ms | Ops: 686.39 ± 2.62 ops/s. Best: 690.52 ops/s<br/>Total: 1457.05ms ± 5.76ms. Best: 1448.17ms<br/>Min time: 1.17ms ± 0.01ms<br/>Mean time: 1.46ms ± 0.01ms<br/>Median time: 1.42ms ± 0.01ms<br/>95th percentile: 1.83ms ± 0.01ms<br/>99th percentile: 1.92ms ± 0.01ms<br/>Max time: 8.90ms ± 7.11ms |
| e2e_nearest_mld | Ops: 679.55 ± 5.01 ops/s. Best: 687.40 ops/s<br/>Total: 1471.76ms ± 11.28ms. Best: 1454.76ms<br/>Min time: 1.21ms ± 0.00ms<br/>Mean time: 1.47ms ± 0.01ms<br/>Median time: 1.43ms ± 0.01ms<br/>95th percentile: 1.82ms ± 0.01ms<br/>99th percentile: 1.91ms ± 0.02ms<br/>Max time: 8.96ms ± 7.19ms | Ops: 671.77 ± 2.49 ops/s. Best: 677.09 ops/s<br/>Total: 1488.54ms ± 5.44ms. Best: 1476.90ms<br/>Min time: 1.22ms ± 0.01ms<br/>Mean time: 1.49ms ± 0.01ms<br/>Median time: 1.45ms ± 0.01ms<br/>95th percentile: 1.84ms ± 0.01ms<br/>99th percentile: 1.92ms ± 0.01ms<br/>Max time: 8.93ms ± 7.07ms |
| e2e_route_ch | Ops: 223.73 ± 0.48 ops/s. Best: 224.44 ops/s<br/>Total: 4469.51ms ± 9.78ms. Best: 4455.52ms<br/>Min time: 1.77ms ± 0.06ms<br/>Mean time: 4.47ms ± 0.01ms<br/>Median time: 4.54ms ± 0.01ms<br/>95th percentile: 5.92ms ± 0.04ms<br/>99th percentile: 6.43ms ± 0.04ms<br/>Max time: 12.87ms ± 5.99ms | Ops: 224.34 ± 0.31 ops/s. Best: 224.87 ops/s<br/>Total: 4457.27ms ± 6.25ms. Best: 4446.93ms<br/>Min time: 1.73ms ± 0.04ms<br/>Mean time: 4.46ms ± 0.01ms<br/>Median time: 4.52ms ± 0.02ms<br/>95th percentile: 5.96ms ± 0.02ms<br/>99th percentile: 6.49ms ± 0.03ms<br/>Max time: 12.83ms ± 5.99ms |
| e2e_route_mld | Ops: 184.34 ± 0.19 ops/s. Best: 184.64 ops/s<br/>Total: 5424.75ms ± 5.55ms. Best: 5415.85ms<br/>Min time: 1.73ms ± 0.03ms<br/>Mean time: 5.42ms ± 0.01ms<br/>Median time: 5.53ms ± 0.01ms<br/>95th percentile: 7.40ms ± 0.01ms<br/>99th percentile: 7.93ms ± 0.06ms<br/>Max time: 14.15ms ± 5.55ms | Ops: 184.43 ± 0.37 ops/s. Best: 185.13 ops/s<br/>Total: 5421.97ms ± 11.31ms. Best: 5401.53ms<br/>Min time: 1.75ms ± 0.04ms<br/>Mean time: 5.42ms ± 0.01ms<br/>Median time: 5.53ms ± 0.02ms<br/>95th percentile: 7.41ms ± 0.02ms<br/>99th percentile: 7.94ms ± 0.05ms<br/>Max time: 14.24ms ± 5.52ms |
| e2e_table_ch | Ops: 219.59 ± 0.52 ops/s. Best: 220.29 ops/s<br/>Total: 4553.79ms ± 10.68ms. Best: 4539.48ms<br/>Min time: 2.49ms ± 0.03ms<br/>Mean time: 4.55ms ± 0.01ms<br/>Median time: 4.55ms ± 0.02ms<br/>95th percentile: 6.21ms ± 0.02ms<br/>99th percentile: 6.53ms ± 0.03ms<br/>Max time: 13.35ms ± 6.78ms | Ops: 220.44 ± 0.39 ops/s. Best: 221.11 ops/s<br/>Total: 4536.31ms ± 8.29ms. Best: 4522.54ms<br/>Min time: 2.46ms ± 0.04ms<br/>Mean time: 4.54ms ± 0.01ms<br/>Median time: 4.55ms ± 0.01ms<br/>95th percentile: 6.19ms ± 0.01ms<br/>99th percentile: 6.50ms ± 0.02ms<br/>Max time: 13.28ms ± 6.77ms |
| e2e_table_mld | Ops: 70.45 ± 0.04 ops/s. Best: 70.50 ops/s<br/>Total: 14194.27ms ± 7.28ms. Best: 14184.70ms<br/>Min time: 5.78ms ± 0.06ms<br/>Mean time: 14.19ms ± 0.01ms<br/>Median time: 14.19ms ± 0.02ms<br/>95th percentile: 21.66ms ± 0.02ms<br/>99th percentile: 22.75ms ± 0.06ms<br/>Max time: 29.05ms ± 5.09ms | Ops: 70.31 ± 0.11 ops/s. Best: 70.46 ops/s<br/>Total: 14222.87ms ± 22.12ms. Best: 14193.28ms<br/>Min time: 5.89ms ± 0.08ms<br/>Mean time: 14.22ms ± 0.02ms<br/>Median time: 14.22ms ± 0.02ms<br/>95th percentile: 21.70ms ± 0.04ms<br/>99th percentile: 22.70ms ± 0.08ms<br/>Max time: 29.15ms ± 5.27ms |
| e2e_trip_ch | Ops: 63.88 ± 0.03 ops/s. Best: 63.94 ops/s<br/>Total: 15655.34ms ± 8.88ms. Best: 15638.76ms<br/>Min time: 2.26ms ± 0.16ms<br/>Mean time: 15.66ms ± 0.01ms<br/>Median time: 14.91ms ± 0.07ms<br/>95th percentile: 27.72ms ± 0.04ms<br/>99th percentile: 29.78ms ± 0.09ms<br/>Max time: 32.23ms ± 1.08ms | Ops: 63.93 ± 0.05 ops/s. Best: 64.01 ops/s<br/>Total: 15641.94ms ± 13.79ms. Best: 15623.74ms<br/>Min time: 2.27ms ± 0.14ms<br/>Mean time: 15.64ms ± 0.01ms<br/>Median time: 14.87ms ± 0.04ms<br/>95th percentile: 27.79ms ± 0.06ms<br/>99th percentile: 29.85ms ± 0.07ms<br/>Max time: 32.41ms ± 1.08ms |
| e2e_trip_mld | Ops: 37.38 ± 0.01 ops/s. Best: 37.40 ops/s<br/>Total: 26754.00ms ± 8.88ms. Best: 26740.86ms<br/>Min time: 2.29ms ± 0.15ms<br/>Mean time: 26.75ms ± 0.01ms<br/>Median time: 25.95ms ± 0.03ms<br/>95th percentile: 43.65ms ± 0.04ms<br/>99th percentile: 46.26ms ± 0.16ms<br/>Max time: 48.59ms ± 0.11ms | Ops: 37.47 ± 0.04 ops/s. Best: 37.52 ops/s<br/>Total: 26687.20ms ± 25.73ms. Best: 26650.22ms<br/>Min time: 2.31ms ± 0.16ms<br/>Mean time: 26.69ms ± 0.03ms<br/>Median time: 25.85ms ± 0.04ms<br/>95th percentile: 43.59ms ± 0.07ms<br/>99th percentile: 46.32ms ± 0.34ms<br/>Max time: 49.56ms ± 1.54ms |
| json-render | String: 8.81307ms<br/>Stringstream: 14.0214ms<br/>Vector: 9.66424ms | String: 8.73489ms<br/>Stringstream: 14.1779ms<br/>Vector: 9.43052ms |
| match_ch | Default radius: <br/>7.06658ms/req at 82 coordinate<br/>0.0861777ms/coordinate<br/>Radius 10m: <br/>24.9958ms/req at 82 coordinate<br/>0.304826ms/coordinate | Default radius: <br/>7.06004ms/req at 82 coordinate<br/>0.086098ms/coordinate<br/>Radius 10m: <br/>25.0429ms/req at 82 coordinate<br/>0.305401ms/coordinate |
| match_mld | Default radius: <br/>4.31222ms/req at 82 coordinate<br/>0.0525881ms/coordinate<br/>Radius 10m: <br/>16.1473ms/req at 82 coordinate<br/>0.196919ms/coordinate | Default radius: <br/>4.33119ms/req at 82 coordinate<br/>0.0528193ms/coordinate<br/>Radius 10m: <br/>16.1721ms/req at 82 coordinate<br/>0.197221ms/coordinate |
| node_match_ch | Ops: 152.5 ± 0.6 ops/s. Best: 153.7 ops/s | Ops: 153.6 ± 0.2 ops/s. Best: 153.9 ops/s |
| node_match_mld | Ops: 217.1 ± 1.2 ops/s. Best: 218.8 ops/s | Ops: 216.2 ± 0.4 ops/s. Best: 216.8 ops/s |
| node_nearest_ch | Ops: 10098.5 ± 835.4 ops/s. Best: 11680.9 ops/s | Ops: 10090.2 ± 696.6 ops/s. Best: 11518.2 ops/s |
| node_nearest_mld | Ops: 10014.6 ± 719.5 ops/s. Best: 11141.0 ops/s | Ops: 9708.8 ± 337.7 ops/s. Best: 10257.3 ops/s |
| node_route_ch | Ops: 1029.2 ± 15.1 ops/s. Best: 1046.6 ops/s | Ops: 998.2 ± 12.0 ops/s. Best: 1015.6 ops/s |
| node_route_mld | Ops: 527.0 ± 2.6 ops/s. Best: 531.1 ops/s | Ops: 528.6 ± 3.7 ops/s. Best: 533.5 ops/s |
| node_table_ch | Ops: 181.6 ± 1.5 ops/s. Best: 183.3 ops/s | Ops: 181.4 ± 1.1 ops/s. Best: 183.4 ops/s |
| node_table_mld | Ops: 38.3 ± 0.0 ops/s. Best: 38.4 ops/s | Ops: 38.5 ± 0.1 ops/s. Best: 38.7 ops/s |
| node_trip_ch | Ops: 185.4 ± 1.4 ops/s. Best: 187.2 ops/s | Ops: 186.6 ± 0.9 ops/s. Best: 187.4 ops/s |
| node_trip_mld | Ops: 62.7 ± 0.1 ops/s. Best: 62.8 ops/s | Ops: 62.9 ± 0.1 ops/s. Best: 63.0 ops/s |
| osrm_contract | Time: 184.91s Peak RAM: 186.04MB | Time: 185.12s Peak RAM: 186.04MB |
| osrm_customize | Time: 2.50s Peak RAM: 112.07MB | Time: 2.48s Peak RAM: 112.07MB |
| osrm_extract | Time: 24.06s Peak RAM: 396.45MB | Time: 24.02s Peak RAM: 403.87MB |
| osrm_partition | Time: 5.61s Peak RAM: 123.48MB | Time: 5.61s Peak RAM: 124.50MB |
| packedvector | random write:<br/>std::vector 174644 ms<br/>util::packed_vector 358117 ms<br/>slowdown: 2.05055<br/>random read:<br/>std::vector 95570.3 ms<br/>util::packed_vector 182176 ms<br/>slowdown: 1.9062 | random write:<br/>std::vector 173784 ms<br/>util::packed_vector 357439 ms<br/>slowdown: 2.0568<br/>random read:<br/>std::vector 94998.4 ms<br/>util::packed_vector 181754 ms<br/>slowdown: 1.91323 |
| random_match_ch | 500 matches, default radius<br/>ops: 123.05 ± 0.18 ops/s. best: 123.24ops/s.<br/>total: 463.24 ± 0.71ms. best: 462.52ms.<br/>avg: 8.13 ± 0.01ms<br/>min: 0.23 ± 0.00ms<br/>max: 41.27 ± 0.04ms<br/>p99: 41.27 ± 0.04ms<br/><br/>500 matches, radius=10<br/>ops: 36.29 ± 0.01 ops/s. best: 36.30ops/s.<br/>total: 1763.62 ± 0.42ms. best: 1763.15ms.<br/>avg: 27.56 ± 0.01ms<br/>min: 0.22 ± 0.00ms<br/>max: 333.06 ± 0.81ms<br/>p99: 333.06 ± 0.81ms<br/><br/>500 matches, radius=20<br/>ops: 8.63 ± 0.02 ops/s. best: 8.66ops/s.<br/>total: 7533.43 ± 14.47ms. best: 7507.30ms.<br/>avg: 115.90 ± 0.22ms<br/>min: 0.48 ± 0.00ms<br/>max: 1866.86 ± 4.73ms<br/>p99: 1866.86 ± 4.73ms<br/><br/>Peak RAM: 56.000MB | 500 matches, default radius<br/>ops: 123.22 ± 0.19 ops/s. best: 123.44ops/s.<br/>total: 462.58 ± 0.73ms. best: 461.77ms.<br/>avg: 8.12 ± 0.01ms<br/>min: 0.23 ± 0.00ms<br/>max: 41.22 ± 0.08ms<br/>p99: 41.22 ± 0.08ms<br/><br/>500 matches, radius=10<br/>ops: 36.43 ± 0.02 ops/s. best: 36.47ops/s.<br/>total: 1756.77 ± 0.90ms. best: 1754.99ms.<br/>avg: 27.45 ± 0.01ms<br/>min: 0.22 ± 0.00ms<br/>max: 330.49 ± 0.85ms<br/>p99: 330.49 ± 0.85ms<br/><br/>500 matches, radius=20<br/>ops: 8.67 ± 0.01 ops/s. best: 8.69ops/s.<br/>total: 7495.28 ± 11.00ms. best: 7476.56ms.<br/>avg: 115.31 ± 0.17ms<br/>min: 0.48 ± 0.00ms<br/>max: 1849.69 ± 5.75ms<br/>p99: 1849.69 ± 5.75ms<br/><br/>Peak RAM: 55.500MB |
| random_match_mld | 500 matches, default radius<br/>ops: 200.91 ± 0.43 ops/s. best: 201.42ops/s.<br/>total: 283.72 ± 0.64ms. best: 282.98ms.<br/>avg: 4.98 ± 0.01ms<br/>min: 0.21 ± 0.00ms<br/>max: 30.27 ± 0.02ms<br/>p99: 30.27 ± 0.02ms<br/><br/>500 matches, radius=10<br/>ops: 74.39 ± 0.09 ops/s. best: 74.53ops/s.<br/>total: 860.33 ± 1.05ms. best: 858.71ms.<br/>avg: 13.44 ± 0.02ms<br/>min: 0.20 ± 0.00ms<br/>max: 137.04 ± 0.48ms<br/>p99: 137.04 ± 0.48ms<br/><br/>500 matches, radius=20<br/>ops: 15.60 ± 0.01 ops/s. best: 15.61ops/s.<br/>total: 4167.45 ± 1.98ms. best: 4164.69ms.<br/>avg: 64.11 ± 0.03ms<br/>min: 0.29 ± 0.00ms<br/>max: 748.78 ± 0.48ms<br/>p99: 748.78 ± 0.48ms<br/><br/>Peak RAM: 51.500MB | 500 matches, default radius<br/>ops: 201.90 ± 0.41 ops/s. best: 202.39ops/s.<br/>total: 282.31 ± 0.63ms. best: 281.63ms.<br/>avg: 4.95 ± 0.01ms<br/>min: 0.21 ± 0.00ms<br/>max: 30.14 ± 0.02ms<br/>p99: 30.14 ± 0.02ms<br/><br/>500 matches, radius=10<br/>ops: 74.81 ± 0.04 ops/s. best: 74.85ops/s.<br/>total: 855.50 ± 0.52ms. best: 855.00ms.<br/>avg: 13.37 ± 0.01ms<br/>min: 0.21 ± 0.00ms<br/>max: 136.30 ± 0.32ms<br/>p99: 136.30 ± 0.32ms<br/><br/>500 matches, radius=20<br/>ops: 15.71 ± 0.00 ops/s. best: 15.72ops/s.<br/>total: 4137.88 ± 1.23ms. best: 4135.57ms.<br/>avg: 63.66 ± 0.02ms<br/>min: 0.29 ± 0.00ms<br/>max: 743.22 ± 1.10ms<br/>p99: 743.22 ± 1.10ms<br/><br/>Peak RAM: 51.500MB |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 21611.88 ± 33.84 ops/s. best: 21636.41ops/s.<br/>total: 462.71 ± 0.73ms. best: 462.18ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.14 ± 0.01ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15595.15 ± 5.63 ops/s. best: 15600.78ops/s.<br/>total: 641.23 ± 0.23ms. best: 640.99ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 12096.76 ± 3.32 ops/s. best: 12103.32ops/s.<br/>total: 826.67 ± 0.23ms. best: 826.22ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.00ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 35.000MB | 10000 nearest, number_of_results=1<br/>ops: 21684.98 ± 30.55 ops/s. best: 21708.19ops/s.<br/>total: 461.15 ± 0.65ms. best: 460.66ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.14 ± 0.01ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15668.48 ± 8.62 ops/s. best: 15677.29ops/s.<br/>total: 638.22 ± 0.35ms. best: 637.87ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 12173.70 ± 8.49 ops/s. best: 12181.76ops/s.<br/>total: 821.44 ± 0.58ms. best: 820.90ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.00ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 35.000MB |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 21643.19 ± 34.63 ops/s. best: 21669.94ops/s.<br/>total: 462.04 ± 0.74ms. best: 461.47ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.14 ± 0.01ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15604.92 ± 12.95 ops/s. best: 15614.58ops/s.<br/>total: 640.82 ± 0.53ms. best: 640.43ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 12099.42 ± 0.67 ops/s. best: 12100.06ops/s.<br/>total: 826.49 ± 0.05ms. best: 826.44ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.00ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 35.000MB | 10000 nearest, number_of_results=1<br/>ops: 21703.06 ± 48.71 ops/s. best: 21740.51ops/s.<br/>total: 460.77 ± 1.04ms. best: 459.97ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.14 ± 0.01ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 15686.04 ± 8.27 ops/s. best: 15692.21ops/s.<br/>total: 637.51 ± 0.34ms. best: 637.26ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 12180.01 ± 5.86 ops/s. best: 12186.48ops/s.<br/>total: 821.02 ± 0.39ms. best: 820.58ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 0.19 ± 0.01ms<br/>p99: 0.15 ± 0.00ms<br/><br/>Peak RAM: 35.000MB |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 289.96 ± 0.57 ops/s. best: 290.42ops/s.<br/>total: 3393.63 ± 6.70ms. best: 3388.22ms.<br/>avg: 3.45 ± 0.01ms<br/>min: 0.51 ± 0.00ms<br/>max: 6.00 ± 0.22ms<br/>p99: 5.15 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 338.01 ± 0.04 ops/s. best: 338.06ops/s.<br/>total: 2958.51 ± 0.37ms. best: 2958.03ms.<br/>avg: 2.96 ± 0.00ms<br/>min: 0.08 ± 0.00ms<br/>max: 6.79 ± 0.04ms<br/>p99: 6.59 ± 0.01ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 600.36 ± 0.11 ops/s. best: 600.53ops/s.<br/>total: 1639.02 ± 0.30ms. best: 1638.55ms.<br/>avg: 1.67 ± 0.00ms<br/>min: 0.37 ± 0.00ms<br/>max: 2.88 ± 0.00ms<br/>p99: 2.41 ± 0.00ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 638.73 ± 0.10 ops/s. best: 638.97ops/s.<br/>total: 1565.61 ± 0.25ms. best: 1565.01ms.<br/>avg: 1.57 ± 0.00ms<br/>min: 0.06 ± 0.00ms<br/>max: 4.14 ± 0.00ms<br/>p99: 4.03 ± 0.00ms<br/><br/>Peak RAM: 84.500MB | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 291.86 ± 0.19 ops/s. best: 292.11ops/s.<br/>total: 3371.43 ± 2.17ms. best: 3368.62ms.<br/>avg: 3.43 ± 0.00ms<br/>min: 0.54 ± 0.00ms<br/>max: 5.89 ± 0.01ms<br/>p99: 5.14 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 337.31 ± 0.09 ops/s. best: 337.52ops/s.<br/>total: 2964.66 ± 0.79ms. best: 2962.79ms.<br/>avg: 2.96 ± 0.00ms<br/>min: 0.08 ± 0.00ms<br/>max: 6.81 ± 0.03ms<br/>p99: 6.65 ± 0.00ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 592.84 ± 0.19 ops/s. best: 593.14ops/s.<br/>total: 1659.81 ± 0.53ms. best: 1658.96ms.<br/>avg: 1.69 ± 0.00ms<br/>min: 0.38 ± 0.00ms<br/>max: 2.92 ± 0.00ms<br/>p99: 2.43 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 630.41 ± 0.08 ops/s. best: 630.50ops/s.<br/>total: 1586.27 ± 0.20ms. best: 1586.05ms.<br/>avg: 1.59 ± 0.00ms<br/>min: 0.06 ± 0.00ms<br/>max: 4.18 ± 0.00ms<br/>p99: 4.07 ± 0.00ms<br/><br/>Peak RAM: 84.000MB |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 148.57 ± 0.05 ops/s. best: 148.63ops/s.<br/>total: 6623.30 ± 2.17ms. best: 6620.67ms.<br/>avg: 6.73 ± 0.00ms<br/>min: 0.54 ± 0.00ms<br/>max: 15.38 ± 0.02ms<br/>p99: 11.02 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 141.15 ± 0.01 ops/s. best: 141.17ops/s.<br/>total: 7084.79 ± 0.65ms. best: 7083.49ms.<br/>avg: 7.08 ± 0.00ms<br/>min: 0.07 ± 0.00ms<br/>max: 16.03 ± 0.07ms<br/>p99: 13.93 ± 0.03ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 206.42 ± 0.01 ops/s. best: 206.44ops/s.<br/>total: 4767.00 ± 0.25ms. best: 4766.52ms.<br/>avg: 4.84 ± 0.00ms<br/>min: 0.39 ± 0.00ms<br/>max: 12.66 ± 0.01ms<br/>p99: 8.23 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 177.10 ± 0.03 ops/s. best: 177.15ops/s.<br/>total: 5646.60 ± 1.04ms. best: 5644.88ms.<br/>avg: 5.65 ± 0.00ms<br/>min: 0.06 ± 0.00ms<br/>max: 12.53 ± 0.02ms<br/>p99: 10.58 ± 0.01ms<br/><br/>Peak RAM: 74.047MB | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 150.48 ± 0.03 ops/s. best: 150.51ops/s.<br/>total: 6539.23 ± 1.41ms. best: 6537.60ms.<br/>avg: 6.65 ± 0.00ms<br/>min: 0.53 ± 0.00ms<br/>max: 15.20 ± 0.01ms<br/>p99: 10.88 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 142.23 ± 0.03 ops/s. best: 142.26ops/s.<br/>total: 7031.10 ± 1.34ms. best: 7029.25ms.<br/>avg: 7.03 ± 0.00ms<br/>min: 0.07 ± 0.00ms<br/>max: 15.99 ± 0.05ms<br/>p99: 13.80 ± 0.02ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 208.64 ± 0.02 ops/s. best: 208.66ops/s.<br/>total: 4716.29 ± 0.39ms. best: 4715.80ms.<br/>avg: 4.79 ± 0.00ms<br/>min: 0.39 ± 0.00ms<br/>max: 12.55 ± 0.01ms<br/>p99: 8.16 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 178.13 ± 0.02 ops/s. best: 178.17ops/s.<br/>total: 5613.95 ± 0.76ms. best: 5612.76ms.<br/>avg: 5.61 ± 0.00ms<br/>min: 0.06 ± 0.00ms<br/>max: 12.46 ± 0.01ms<br/>p99: 10.51 ± 0.00ms<br/><br/>Peak RAM: 73.984MB |
…